+++ src/Quests/ICompletionArgs.cs
@@ -28,5 +28,10 @@ namespace QuestFramework.Quests
         /// Text to test
         /// </summary>
         string String { get; }
+
+        /// <summary>
+        /// Type of completion source (Quest type id)
+        /// </summary>
+        int CompletionType { get; }
     }
 }